refactor: :truck: update utils folder name to helpers
