+++ src/components/CustomHeader/CustomHeader.test.tsx
@@ -1,6 +1,6 @@
 import { render, screen } from '@testing-library/react'
 import CustomHeader from './CustomHeader'
-import { navLinks } from './navLinks'
+import { navLinks } from '../../helpers/navLinks'
 
 describe('Welcome component', () => {
   it('has correct Next.js theming section link', () => {
